split server into script and package.
